+++ bedrock/src/server/patterns.js
@@ -17,8 +17,9 @@
 const { gql } = require('apollo-server-express');
 const GraphQLJSON = require('graphql-type-json');
 const fs = require('fs-extra');
-const { join, relative } = require('path');
+const { join, relative, resolve, parse } = require('path');
 const globby = require('globby');
+const { getFileSizes } = require('get-file-sizes');
 const {
   validateSchemaAndAssignDefaults,
   validateUniqueIdsInArray,
@@ -27,11 +28,12 @@ const chokidar = require('chokidar');
 const {
   version: iframeResizerVersion,
 } = require('iframe-resizer/package.json');
+const qs = require('qs');
 const { bedrockEvents, EVENTS } = require('./events');
 const { FileDb } = require('./db');
 const patternSchema = require('../schemas/pattern.schema');
 const patternMetaSchema = require('../schemas/pattern-meta.schema');
-const { writeJson, fileExists } = require('./server-utils');
+const { writeJson, fileExists, fileExistsOrExit } = require('./server-utils');
 const log = require('../cli/log');
 const { FILE_NAMES } = require('../lib/constants');
 
@@ -51,6 +53,19 @@ const patternsTypeDef = gql`
     items: [PatternDoAndDontItem!]!
   }
 
+  type PatternAsset {
+    src: String!
+    publicPath: String!
+    sizeRaw: String
+    sizeKb: String
+  }
+
+  type PatternAssetSet {
+    id: ID!
+    title: String!
+    assets: [PatternAsset]
+  }
+
   type PatternTemplate {
     "JSON Schema"
     schema: JSON
@@ -62,9 +77,12 @@ const patternsTypeDef = gql`
     docPath: String
     doc: String
     demoDatas: [JSON]
+    demoUrls: [String]
     uiSchema: JSON
     isInline: Boolean
     demoSize: String
+    assetSets: [PatternAssetSet]
+    hideCodeBlock: Boolean
   }
 
   type PatternType {
@@ -123,9 +141,22 @@ const patternsTypeDef = gql`
     message: String
   }
 
+  type TemplateCode {
+    usage: String
+    data: JSON
+    templateSrc: String
+    html: String
+    language: String
+  }
+
   type Query {
     patterns: [Pattern]
     pattern(id: ID): Pattern
+    templateCode(
+      patternId: String
+      templateId: String
+      data: JSON
+    ): TemplateCode
     patternTypes: [PatternType]
     patternType(id: ID): PatternType
     patternStatuses: [PatternStatus]
@@ -248,11 +279,65 @@ async function writeAllFiles(dir, config) {
 }
 
 /**
+ * @param {string} url
+ * @return {boolean}
+ */
+function isRemoteUrl(url) {
+  return url.startsWith('http') || url.startsWith('//');
+}
+
+/**
+ * @param {Object} opt
+ * @param {BedrockAssetSetUserConfig[]} opt.assetSets
+ * @param {string} opt.publicDir
+ * @param {string} opt.configPathBase
+ * @return {BedrockAssetSet[]}
+ */
+function processAssetSets({ assetSets, publicDir, configPathBase }) {
+  return assetSets.map(assetSet => ({
+    ...assetSet,
+    assets: assetSet.assets.map(asset => {
+      const isRemote = isRemoteUrl(asset.src);
+      const src = isRemote ? asset.src : resolve(configPathBase, asset.src);
+
+      if (!isRemote) {
+        fileExistsOrExit(src);
+        if (relative(publicDir, src).includes('..')) {
+          log.error(
+            `Some CSS or JS is not publically accessible! These must be either remote or places inside the "public" dir (${publicDir})`,
+          );
+          process.exit(1);
+        }
+      }
+
+      const { ext } = parse(src);
+      const [size] = getFileSizes([src]);
+
+      return {
+        src,
+        // isInHead: asset.isInHead === true,
+        publicPath: isRemote ? src : `/${relative(publicDir, src)}`,
+        type: ext.replace('.', ''),
+        sizeRaw: size.sizeRaw,
+        sizeKb: size.sizeKb,
+      };
+    }),
+  }));
+}
+
+/**
  * @param {string[]} patternsDirs
  * @param {BedrockTemplateRenderer[]} templateRenderers
+ * @param {function(BedrockAssetSetUserConfig[]): BedrockAssetSet[]} scanAssetSets
+ * @param {BedrockAssetSet[]} globalAssetSets
  * @returns {BedrockPattern[]}
  */
-function createPatternsData(patternsDirs, templateRenderers) {
+function createPatternsData(
+  patternsDirs,
+  templateRenderers,
+  scanAssetSets,
+  globalAssetSets,
+) {
   const patterns = [];
 
   patternsDirs.forEach(dir => {
@@ -336,10 +421,58 @@ function createPatternsData(patternsDirs, templateRenderers) {
             doc = fs.readFileSync(docPath, 'utf8');
           }
 
+          const { schema, demoDatas, id: templateId } = template;
+
+          const hasSchema = !!(
+            schema &&
+            schema.properties &&
+            Object.keys(schema.properties).length > 0
+          );
+
+          let datas = [{}];
+          if (demoDatas) {
+            datas = demoDatas;
+          } else if (
+            hasSchema &&
+            schema.examples &&
+            schema.examples.length > 0
+          ) {
+            datas = schema.examples;
+          }
+
+          const demoUrls = datas.map(data => {
+            const queryString = qs.stringify({
+              templateId,
+              patternId: pattern.id,
+              data,
+              isInIframe: false,
+              wrapHtml: true,
+            });
+
+            return `/api/render?${queryString}`;
+          });
+
+          let assetSets = [];
+          if (template.assetSets) {
+            assetSets = scanAssetSets(template.assetSets);
+          }
+
+          globalAssetSets.forEach(globalAssetSet => {
+            if (!assetSets.some(a => a.id === globalAssetSet.id)) {
+              assetSets.push(globalAssetSet);
+            }
+          });
+
+          const src = fs.readFileSync(templatePath, 'utf8');
+
           return {
             ...template,
+            demoDatas: datas,
+            demoUrls,
             absolutePath: templatePath,
+            assetSets,
             doc,
+            src,
           };
         });
 
@@ -420,9 +553,10 @@ class Patterns {
    * @param {Object} opt
    * @param {string} opt.newPatternDir
    * @param {string[]} opt.patternPaths
+   * @param {string} opt.publicDir
+   * @param {string} opt.configPathBase
    * @param {string} opt.dataDir
-   * @param {string[]} opt.rootRelativeJs
-   * @param {string[]} opt.rootRelativeCSS
+   * @param {BedrockAssetSetUserConfig[]} opt.assetSets
    * @param {BedrockTemplateRenderer[]} opt.templateRenderers
    */
   constructor({
@@ -430,8 +564,9 @@ class Patterns {
     patternPaths,
     dataDir,
     templateRenderers,
-    rootRelativeJs,
-    rootRelativeCSS,
+    assetSets,
+    publicDir,
+    configPathBase,
   }) {
     this.db = new FileDb({
       dbDir: dataDir,
@@ -463,6 +598,20 @@ class Patterns {
       },
     });
 
+    /**
+     * @param {BedrockAssetSetUserConfig[]} newAssetSets
+     * @return {BedrockAssetSet[]}
+     */
+    this.scanAssetSets = newAssetSets =>
+      processAssetSets({
+        assetSets: newAssetSets,
+        publicDir,
+        configPathBase,
+      });
+
+    /** @type {BedrockAssetSet[]} */
+    this.globalAssetSets = this.scanAssetSets(assetSets);
+
     /** @type {string} */
     this.newPatternDir = newPatternDir;
     /** @type {string[]} */
@@ -471,10 +620,6 @@ class Patterns {
     this.dataDir = dataDir;
 
     this.templateRenderers = templateRenderers;
-    /** @type {string[]} */
-    this.rootRelativeCSS = rootRelativeCSS;
-    /** @type {string[]} */
-    this.rootRelativeJs = rootRelativeJs;
 
     /** @type {string[]} */
     this.patternsDirs = getPatternsDirs(this.patternPaths);
@@ -483,6 +628,8 @@ class Patterns {
     this.allPatterns = createPatternsData(
       this.patternsDirs,
       this.templateRenderers,
+      this.scanAssetSets,
+      this.globalAssetSets,
     );
   }
 
@@ -491,6 +638,8 @@ class Patterns {
     this.allPatterns = createPatternsData(
       this.patternsDirs,
       this.templateRenderers,
+      this.scanAssetSets,
+      this.globalAssetSets,
     );
   }
 
@@ -555,6 +704,26 @@ class Patterns {
       });
     });
     return allTemplatePaths;
+  }
+
+  getAllAssetPaths(includeRemote = false) {
+    const paths = new Set();
+    const patterns = this.getPatterns();
+    patterns.forEach(pattern => {
+      pattern.templates.forEach(template => {
+        template.assetSets.forEach(({ assets }) => {
+          assets.forEach(asset => {
+            if (includeRemote) {
+              paths.add(asset.src);
+            } else if (!isRemoteUrl(asset.src)) {
+              paths.add(asset.src);
+            }
+          });
+        });
+      });
+    });
+
+    return [...paths];
   }
 
   /**
@@ -648,6 +817,96 @@ class Patterns {
     this.db.setAll(settings);
   }
 
+  /**
+   * @param {string} patternId
+   * @return {{ id: string, title: string, demoUrls: string[] }[]}
+   */
+  getPatternDemoUrls(patternId) {
+    const pattern = this.getPattern(patternId);
+
+    return pattern.templates.map(template => ({
+      id: template.id,
+      title: template.title,
+      demoUrls: template.demoUrls,
+    }));
+  }
+
+  /**
+   * @return {{templates: {id: string, title: string, demoUrls: string[]}[], id: Id, title: Title}[]}
+   */
+  getPatternsDemoUrls() {
+    return this.getPatterns().map(pattern => ({
+      id: pattern.id,
+      title: pattern.meta.title,
+      templates: this.getPatternDemoUrls(pattern.id),
+    }));
+  }
+
+  /**
+   * Get code strings to help with how this template is used
+   * @param {Object} opt
+   * @param {string} opt.patternId
+   * @param {string} opt.templateId
+   * @param {Object} [opt.data] - data passed to template
+   * @return {Promise<BedrockPatternTemplateCode>}
+   */
+  async getTemplateCode({ patternId, templateId, data }) {
+    const pattern = this.getPattern(patternId);
+    const template = pattern.templates.find(t => t.id === templateId);
+    const renderer = this.templateRenderers.find(t =>
+      t.test(template.absolutePath),
+    );
+
+    let language = 'bash';
+    if (renderer.language) {
+      language = renderer.language; // eslint-disable-line
+    } else if (renderer.extension) {
+      language = renderer.extension.replace('.', '');
+    }
+
+    const results = await Promise.all([
+      renderer
+        .getUsage({ patternId, template, data })
+        .then(usage => ({ usage })),
+      this.render({
+        patternId,
+        templateId,
+        data,
+        wrapHtml: false,
+        isInIframe: false,
+      }).then(({ ok, html, message }) => {
+        if (!ok) {
+          log.error(`Error trying to getTemplateCode(): ${message}`, {
+            patternId,
+            templateId,
+            html,
+            message,
+          });
+          throw new Error(message);
+        }
+        return {
+          html,
+        };
+      }),
+    ]);
+
+    const mergedResults = results.reduce(
+      (prev, current) => Object.assign(prev, current),
+      {},
+    );
+    return Object.assign(
+      {},
+      {
+        language,
+        templateSrc: template.src,
+        html: '',
+        usage: '',
+        data,
+      },
+      mergedResults,
+    );
+  }
+
   watch() {
     const configFilesToWatch = [];
     this.allPatterns.forEach(pattern => {
@@ -675,6 +934,16 @@ class Patterns {
       bedrockEvents.emit(EVENTS.PATTERN_CONFIG_CHANGED, { event, path });
       this.updatePatternsData();
     });
+
+    const localAssetPaths = this.getAllAssetPaths(false);
+
+    const assetWatcher = chokidar.watch(localAssetPaths, {
+      ignoreInitial: true,
+    });
+
+    assetWatcher.on('all', (event, path) => {
+      bedrockEvents.emit(EVENTS.PATTERN_ASSET_CHANGED, { event, path });
+    });
   }
 
   /**
@@ -685,6 +954,7 @@ class Patterns {
    * @param {boolean} [opt.wrapHtml=true] - Should it wrap HTML results with `<head>` and include assets?
    * @param {Object} [opt.data] - Data to pass to template
    * @param {boolean} [opt.isInIframe=false] - Will this be in an iFrame?
+   * @param {string} [opt.assetSetId] - Asset Set Id
    * @param {number} [opt.websocketsPort]
    * @return {Promise<BedrockTemplateRenderResults>}
    */
@@ -695,6 +965,7 @@ class Patterns {
     data = {},
     isInIframe = false,
     websocketsPort,
+    assetSetId,
   }) {
     const pattern = this.getPattern(patternId);
     let [template] = pattern.templates;
@@ -713,11 +984,24 @@ class Patterns {
     });
 
     if (!renderedTemplate.ok) return renderedTemplate;
+
     if (wrapHtml) {
-      const inlineJS = [];
+      const assetSet = assetSetId
+        ? template.assetSets.find(a => a.id === assetSetId)
+        : template.assetSets[0];
+
+      const {
+        assets,
+        inlineJs = '',
+        inlineCss = '',
+        inlineFoot = '',
+        inlineHead = '',
+      } = assetSet;
+
+      const inlineJSs = [inlineJs];
 
       if (isInIframe) {
-        inlineJS.push(`
+        inlineJSs.push(`
 /**
   * Prevents the natural click behavior of any links within the iframe.
   * Otherwise the iframe reloads with the current page or follows the url provided.
@@ -730,7 +1014,7 @@ links.forEach(function(link) {
       }
 
       if (!isInIframe && websocketsPort) {
-        inlineJS.push(`
+        inlineJSs.push(`
 if ('WebSocket' in window && location.hostname === 'localhost') {
   var socket = new window.WebSocket('ws://localhost:8000');
   socket.addEventListener('message', function() {
@@ -746,9 +1030,16 @@ if ('WebSocket' in window && location.hostname === 'localhost') {
             ? `https://cdnjs.cloudflare.com/ajax/libs/iframe-resizer/${iframeResizerVersion}/iframeResizer.contentWindow.min.js`
             : '',
         ].filter(x => x),
-        cssUrls: this.rootRelativeCSS,
-        jsUrls: this.rootRelativeJs,
-        inlineJs: inlineJS.join('\n'),
+        cssUrls: assets
+          .filter(asset => asset.type === 'css')
+          .map(asset => asset.publicPath),
+        jsUrls: assets
+          .filter(asset => asset.type === 'js')
+          .map(asset => asset.publicPath),
+        inlineJs: inlineJSs.join('\n'),
+        inlineCss,
+        inlineHead,
+        inlineFoot,
       });
       return {
         ...renderedTemplate,
@@ -763,6 +1054,11 @@ const patternsResolvers = {
   Query: {
     patterns: (parent, args, { patterns }) => patterns.getPatterns(),
     pattern: (parent, { id }, { patterns }) => patterns.getPattern(id),
+    templateCode: async (
+      parent,
+      { patternId, templateId, data },
+      { patterns },
+    ) => patterns.getTemplateCode({ patternId, templateId, data }),
     patternTypes: (parent, args, { patterns }) => patterns.getPatternTypes(),
     patternType: (parent, { id }, { patterns }) => patterns.getPatternType(id),
     patternStatuses: (parent, args, { patterns }) =>