+++ packages/renderer-react/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.39.0](https://github.com/basaltinc/bedrock/compare/v0.38.3...v0.39.0) (2019-03-07)
+
+**Note:** Version bump only for package @basalt/bedrock-renderer-react
+
+
+
+
+
+## [0.38.2](https://github.com/basaltinc/bedrock/compare/v0.38.1...v0.38.2) (2019-03-06)
+
+**Note:** Version bump only for package @basalt/bedrock-renderer-react
+
+
+
+
+
+## [0.38.1](https://github.com/basaltinc/bedrock/compare/v0.38.0...v0.38.1) (2019-03-06)
+
+**Note:** Version bump only for package @basalt/bedrock-renderer-react
+
+
+
+
+
+# [0.38.0](https://github.com/basaltinc/bedrock/compare/v0.37.2...v0.38.0) (2019-03-06)
+
+
+### Features
+
+* add template usage code block ([a8fe90e](https://github.com/basaltinc/bedrock/commit/a8fe90e))
+
+
+
+
+
+# [0.37.0](https://github.com/basaltinc/bedrock/compare/v0.36.0...v0.37.0) (2019-02-24)
+
+**Note:** Version bump only for package @basalt/bedrock-renderer-react
+
+
+
+
+
+# [0.36.0](https://github.com/basaltinc/bedrock/compare/v0.35.0...v0.36.0) (2019-02-23)
+
+**Note:** Version bump only for package @basalt/bedrock-renderer-react
+
+
+
+
+
+# [0.35.0](https://github.com/basaltinc/bedrock/compare/v0.34.2...v0.35.0) (2019-02-23)
+
+**Note:** Version bump only for package @basalt/bedrock-renderer-react
+
+
+
+
+
+# [0.34.0](https://github.com/basaltinc/bedrock/compare/v0.33.3...v0.34.0) (2019-02-17)
+
+**Note:** Version bump only for package @basalt/bedrock-renderer-react
+
+
+
+
+
 # [0.33.0](https://github.com/basaltinc/bedrock/compare/v0.32.5...v0.33.0) (2019-02-15)
 
 **Note:** Version bump only for package @basalt/bedrock-renderer-react